Move component to base
