+++ assets/js/base/components/price-slider/handles.sketch
