+++ assets/js/base/components/price-slider/index.js
