+++ assets/js/base/components/price-slider/style.scss
