+++ assets/js/base/components/price-slider/utils.js
